refactor: separar monolito en Blueprints (routes/api) y Mixins (bd)
Refactorizacion completa de la arquitectura del proyecto sin romper
funcionalidad existente. Se divide el monolito main.py (~736 lineas)
y API.py (~1466 lineas) en modulos independientes.

=== Base de datos (bd/) ===

- Extraer metodos de BDConector a 5 mixins independientes:
  - bd/mixins/users.py: user_exists, add_user, get_user_by_email
  - bd/mixins/items.py: add_item, get_item_by_barcode, get_item_stock,
    total_items, get_item_details, get_item_status, disable_item, enable_item
  - bd/mixins/sales.py: get_dashboard_stats, record_product_sale, record_bulk_sale
  - bd/mixins/metrics.py: get_metrics_data
  - bd/mixins/password_reset.py: save_reset_code, get_reset_code,
    delete_reset_code, update_user_password, verify_code
- BDConector hereda de todos los mixins via herencia multiple
- La interfaz publica (db.metodo()) no cambia: retrocompatibilidad total
- Limpiar docstrings redundantes en bdConector.py

=== API (api/) ===

- Eliminar api/API.py monolitico (~1466 lineas)
- Crear api/__init__.py con Blueprint padre que registra sub-blueprints
- Crear api/auth_utils.py: require_auth() y require_admin() compartidos
- Crear api/health_api.py: endpoint /health
- Crear api/products_api.py: CRUD de productos + busqueda
- Crear api/sales_api.py: registro de ventas + listado + detalle
- Crear api/users_api.py: CRUD usuarios + reset password + actividad
- Crear api/metrics_api.py: estadisticas dashboard + metricas analytics

=== Rutas (routes/) ===

- Crear routes/__init__.py con lista all_blueprints
- Crear routes/auth.py: login, register, logout, reset-password
- Crear routes/dashboard.py: pagina principal
- Crear routes/products.py: formulario producto, gestion, importacion CSV
- Crear routes/sales.py: formulario venta, historial
- Crear routes/settings.py: perfil, cambio de contrasena
- Crear routes/users.py: pagina de gestion de usuarios
- Crear routes/metrics.py: pagina de metricas

=== main.py ===

- Reducir de ~736 a ~80 lineas
- Solo contiene configuracion Flask, registro de blueprints,
  error handlers, signal handlers y arranque webview

=== Templates ===

- Actualizar url_for() en todos los templates para usar namespaced
  blueprints:
  - url_for('index') -> url_for('dashboard.index')
  - url_for('login') -> url_for('auth.login')
  - url_for('logout') -> url_for('auth.logout')
  - url_for('product_new') -> url_for('products.product_new')
  - url_for('import_preview') -> url_for('products.import_preview')
  - url_for('sales') -> url_for('sales.sales')
  - url_for('metrics') -> url_for('metrics.metrics')

  Archivos modificados: base.html, 404.html, dashboard.html,
  import.html, product_management.html, reset_password.html,
  sales.html, settings.html, under_development.html
