--- api/API.py
@@ -1,1466 +1,0 @@
-from flask import Blueprint, jsonify, request, session
-from bd.bdConector import BDConector
-from datetime import datetime, timedelta
-from bd.bdInstance import *
-from tools.logger import logger
-from tools.email import email_sender
-from data.validators import ItemValidator, UserValidator, ValidationError
-from werkzeug.security import generate_password_hash
-
-api_bp = Blueprint("api", __name__)
-
-def require_auth():
-    """
-    Verifica si el usuario está autenticado.
-    Retorna un error JSON si no está autenticado.
-    
-    Requiere login: True.
-    
-    Returns:
-        None si está autenticado, o una respuesta JSON de error si no lo está.
-    """
-    
-    user_id = session.get("user_id")
-    logger.info(f"require_auth check: user_id={user_id}, cookies={dict(request.cookies)}")
-    if not user_id:
-        return jsonify({"error": "Unauthorized"}), 401
-    return None
-
-@api_bp.route("/health", methods=["GET"])
-def health():
-    """
-    Endpoint de verificación de salud del servidor.
-    
-    Requiere login: False.
-    
-    Returns:
-        JSON: {"status": "Ok"} con código 200
-    """
-    return jsonify({"status": "Ok"}), 200
-
-@api_bp.route("/products_all", methods=["GET"])
-def get_all_products():
-    """
-    Obtiene productos del inventario con paginacion server-side.
-
-    Requiere login: True.
-
-    Query Parameters:
-        search      (str, optional) : Busqueda por nombre, codigo de barras o descripcion
-        view_mode   (str, optional) : "all" | "in_stock" | "low_stock" | "out_of_stock"
-        sort        (str, optional) : Campo de orden: "name" | "stock" | "price"  (default: "name")
-        order       (str, optional) : "asc" | "desc"  (default: "asc")
-        page        (int, optional) : Numero de pagina, 1-based  (default: 1)
-        limit       (int, optional) : Registros por pagina, max 250  (default: 50)
-
-    Returns:
-        JSON:
-            data   (list) : Productos de la pagina actual
-            total  (int)  : Total de registros que coinciden con los filtros
-            page   (int)  : Pagina actual
-            pages  (int)  : Total de paginas
-            limit  (int)  : Registros por pagina usados
-
-    Status Codes:
-        200: Exito
-        401: No autorizado
-    """
-
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-
-    # ── Parametros ────────────────────────────────────────────
-    search    = request.args.get("search", "").strip()
-    view_mode = request.args.get("view_mode", "all")
-    sort      = request.args.get("sort", "name")
-    order     = request.args.get("order", "asc").lower()
-    page      = max(1, request.args.get("page", 1, type=int))
-    limit     = min(250, max(1, request.args.get("limit", 50, type=int)))
-    offset    = (page - 1) * limit
-
-    # Whitelist para evitar SQL injection en ORDER BY
-    allowed_sort  = {"name", "stock", "price"}
-    allowed_order = {"asc", "desc"}
-    sort  = sort  if sort  in allowed_sort  else "name"
-    order = order if order in allowed_order else "asc"
-
-    # Mapeo campo logico -> columna real
-    sort_column = {"name": "name", "stock": "quantity", "price": "price"}[sort]
-
-    # ── Filtros WHERE ─────────────────────────────────────────
-    where  = ["1=1"]
-    params = []
-
-    if search:
-        where.append("(name LIKE ? OR barrs_code LIKE ? OR description LIKE ?)")
-        params.extend([f"%{search}%", f"%{search}%", f"%{search}%"])
-
-    if view_mode == "in_stock":
-        where.append("quantity > min_quantity")
-    elif view_mode == "low_stock":
-        where.append("quantity > 0 AND quantity <= min_quantity")
-    elif view_mode == "out_of_stock":
-        where.append("quantity = 0")
-
-    where_clause = " AND ".join(where)
-
-    # ── COUNT total ───────────────────────────────────────────
-    count_query = f"SELECT COUNT(*) FROM items WHERE {where_clause}"
-    total = db.execute_query(count_query, tuple(params))[0][0]
-    pages = max(1, -(-total // limit))  # ceil division
-
-    # ── Pagina actual ─────────────────────────────────────────
-    data_query = f"""
-        SELECT id, barrs_code, name, description, quantity, min_quantity, price, status
-        FROM items
-        WHERE {where_clause}
-        ORDER BY {sort_column} {order}
-        LIMIT ? OFFSET ?
-    """
-    rows = db.execute_query(data_query, tuple(params) + (limit, offset))
-
-    products = [
-        {
-            "id":          row[0],
-            "barcode":     row[1],
-            "name":        row[2],
-            "description": row[3],
-            "stock":       row[4],
-            "min_stock":   row[5],
-            "price":       row[6],
-            "status":      row[7],
-        }
-        for row in rows
-    ]
-
-    return jsonify({
-        "data":  products,
-        "total": total,
-        "page":  page,
-        "pages": pages,
-        "limit": limit,
-    }), 200
-
-@api_bp.route("/products", methods=["GET"])
-def get_products():
-    """
-    Obtiene productos activos del inventario con paginacion server-side.
-
-    Requiere login: True.
-
-    Query Parameters:
-        search      (str, optional) : Busqueda por nombre o codigo de barras
-        view_mode   (str, optional) : "all" | "in_stock" | "out_of_stock"
-        sort        (str, optional) : "name" | "stock" | "price"  (default: "name")
-        order       (str, optional) : "asc" | "desc"  (default: "asc")
-        page        (int, optional) : Pagina, 1-based  (default: 1)
-        limit       (int, optional) : Registros por pagina, max 250  (default: 24)
-
-    Returns:
-        JSON:
-            data   (list) : Productos de la pagina actual
-            total  (int)  : Total de registros con los filtros
-            page   (int)  : Pagina actual
-            pages  (int)  : Total de paginas
-            limit  (int)  : Registros por pagina usados
-
-    Status Codes:
-        200: Exito
-        401: No autorizado
-    """
-
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-
-    search    = request.args.get("search", "").strip()
-    view_mode = request.args.get("view_mode", "all")
-    sort      = request.args.get("sort", "name")
-    order     = request.args.get("order", "asc").lower()
-    page      = max(1, request.args.get("page", 1, type=int))
-    limit     = min(250, max(1, request.args.get("limit", 24, type=int)))
-    offset    = (page - 1) * limit
-
-    allowed_sort  = {"name", "stock", "price"}
-    allowed_order = {"asc", "desc"}
-    sort  = sort  if sort  in allowed_sort  else "name"
-    order = order if order in allowed_order else "asc"
-    sort_column = {"name": "name", "stock": "quantity", "price": "price"}[sort]
-
-    where  = ["status = 1"]
-    params = []
-
-    if search:
-        where.append("(name LIKE ? OR barrs_code LIKE ?)")
-        params.extend([f"%{search}%", f"%{search}%"])
-
-    if view_mode == "in_stock":
-        where.append("quantity > 0")
-    elif view_mode == "out_of_stock":
-        where.append("quantity = 0")
-
-    where_clause = " AND ".join(where)
-
-    total = db.execute_query(
-        f"SELECT COUNT(*) FROM items WHERE {where_clause}", tuple(params)
-    )[0][0]
-    pages = max(1, -(-total // limit))
-
-    rows = db.execute_query(
-        f"""
-        SELECT id, barrs_code, name, description, quantity, min_quantity, price, status
-        FROM items
-        WHERE {where_clause}
-        ORDER BY {sort_column} {order}
-        LIMIT ? OFFSET ?
-        """,
-        tuple(params) + (limit, offset),
-    )
-
-    products = [
-        {
-            "id":          row[0],
-            "barcode":     row[1],
-            "name":        row[2],
-            "description": row[3],
-            "stock":       row[4],
-            "min_stock":   row[5],
-            "price":       row[6],
-            "status":      row[7],
-        }
-        for row in rows
-    ]
-
-    return jsonify({
-        "data":  products,
-        "total": total,
-        "page":  page,
-        "pages": pages,
-        "limit": limit,
-    }), 200
-
-@api_bp.route("/products/<int:product_id>", methods=["GET"])
-def get_product(product_id):
-    """
-    Obtiene un producto específico por su ID (No muestra los deshabilitados).
-    
-    Requiere login: True.
-    
-    Args:
-        product_id (int): ID del producto a buscar
-    
-    Returns:
-        JSON: Detalles completos del producto
-        - id (int): ID del producto
-        - barcode (str): Código de barras
-        - name (str): Nombre
-        - description (str): Descripción
-        - stock (int): Cantidad disponible
-        - min_stock (int): Stock mínimo
-        - price (float): Precio
-    
-    Status Codes:
-        200: Producto encontrado
-        401: No autorizado
-        404: Producto no encontrado
-    """
-    
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-    
-    rows = db.execute_query(
-        "SELECT id, barrs_code, name, description, quantity, min_quantity, price FROM items WHERE id = ?",
-        (product_id,)
-    )
-    
-    if not rows:
-        return jsonify({"error": "Producto no encontrado"}), 404
-    
-    row = rows[0]
-    product = {
-        "id": row[0],
-        "barcode": row[1],
-        "name": row[2],
-        "description": row[3],
-        "stock": row[4],
-        "min_stock": row[5],
-        "price": row[6]
-    }
-    
-    return jsonify(product), 200
-
-@api_bp.route("/products", methods=["POST"])
-def create_product():
-    """
-    Crea un nuevo producto en el inventario.
-    
-    Requiere login: True.
-    Requiere rol: admin.
-    
-    Request Body (JSON):
-        barcode (str): Código de barras del producto
-        name (str): Nombre del producto
-        description (str, optional): Descripción detallada
-        quantity (int): Cantidad inicial en stock
-        min_quantity (int): Stock mínimo de alerta
-        price (float): Precio de venta
-    
-    Returns:
-        JSON: {"message": "Producto creado exitosamente"}
-    
-    Status Codes:
-        201: Producto creado exitosamente
-        400: Faltan campos requeridos
-        401: No autorizado
-        403: Permiso denegado (no es admin)
-        500: Error en la base de datos
-    """
-    
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-    
-    if session.get("role") != "admin":
-        return jsonify({"error": "Permiso denegado"}), 403
-    
-    data = request.get_json()
-    
-    required_fields = ["barcode", "name", "quantity", "min_quantity", "price"]
-    if not all(field in data for field in required_fields):
-        return jsonify({"error": "Faltan campos requeridos"}), 400
-    
-    try:
-        data = ItemValidator.validate(data.get("barcode", ""), data.get("description", ""), data.get("name", ""), 
-                                    data.get("quantity"), data.get("min_quantity"), 
-                                    data.get("price"), 1)
-    except ValidationError as e:
-        return jsonify({"error": e.field + ": " + e.message}), 400
-    
-    try:
-        db.add_item(
-            data.get("barrs_code", ""),
-            data.get("description", ""),
-            data["name"],
-            data["quantity"],
-            data["min_quantity"],
-            data["price"]
-        )
-        return jsonify({"message": "Producto creado exitosamente"}), 201
-    
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
-@api_bp.route("/products/<int:product_id>", methods=["PUT"])
-def update_product(product_id):
-    """
-    Actualiza un producto existente.
-    
-    Requiere login: True.
-    Requiere rol: admin.
-    
-    Args:
-        product_id (int): ID del producto a actualizar
-    
-    Request Body (JSON):
-        name (str, optional): Nuevo nombre
-        description (str, optional): Nueva descripción
-        quantity (int, optional): Nueva cantidad en stock
-        min_quantity (int, optional): Nuevo stock mínimo
-        price (float, optional): Nuevo precio
-        status (int, optional): Nuevo estado (1=activo, 0=deshabilitado)
-    
-    Returns:
-        JSON: {"message": "Producto actualizado"}
-    
-    Status Codes:
-        200: Producto actualizado exitosamente
-        400: No hay datos para actualizar
-        401: No autorizado
-        403: Permiso denegado (no es admin)
-    """
-    
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-    
-    if session.get("role") != "admin":
-        return jsonify({"error": "Permiso denegado"}), 403
-    
-    data = request.get_json()
-    
-    updates = []
-    params = []
-    
-    field_mapping = {
-        "name": "name",
-        "description": "description",
-        "quantity": "quantity",
-        "min_quantity": "min_quantity",
-        "price": "price",
-        "status": "status"
-    }
-    try:
-        data = ItemValidator.validate("0", data.get("description", ""), data.get("name", ""), 
-                                    data.get("quantity"), data.get("min_quantity"), 
-                                    data.get("price"), data.get("status"))
-    except ValidationError as e:
-        return jsonify({"error": e.field + ": " + e.message}), 400
-    
-    for key, db_field in field_mapping.items():
-        if key in data:
-            updates.append(f"{db_field} = ?")
-            params.append(data[key])
-    
-    if not updates:
-        return jsonify({"error": "No hay datos para actualizar"}), 400
-    
-    params.append(product_id)
-    query = f"UPDATE items SET {', '.join(updates)} WHERE id = ?"
-    
-    db.execute_query(query, tuple(params), fetch=False)
-    return jsonify({"message": "Producto actualizado"}), 200
-
-@api_bp.route("/products/<int:product_id>", methods=["DELETE"])
-def delete_product(product_id):
-    """
-    Elimina un producto del inventario.
-    
-    Requiere login: True.
-    Requiere rol: admin.
-    
-    Args:
-        product_id (int): ID del producto a eliminar
-    
-    Returns:
-        JSON: {"message": "Producto eliminado"}
-    
-    Status Codes:
-        200: Producto eliminado exitosamente
-        401: No autorizado
-        403: Permiso denegado (no es admin)
-    """
-    
-    auth_error = require_auth()
-    if auth_error:
-        logger.error(f"Unauthorized delete attempt for product ID {product_id}")
-        return auth_error
-    
-    if session.get("role") != "admin":
-        logger.warning(f"Forbidden delete attempt for product ID {product_id} by user ID {session.get('user_id')}")
-        return jsonify({"error": "Permiso denegado"}), 403
-    
-    db.disable_item(product_id)
-    return jsonify({"message": "Producto eliminado"}), 200
-
-@api_bp.route("/stats", methods=["GET"])
-def get_stats():
-    """
-    Obtiene estadísticas del dashboard.
-    
-    Requiere login: True.
-    
-    Returns:
-        JSON: Estadísticas del sistema
-        - products (int): Total de productos
-        - low_stock (int): Productos con stock bajo
-        - sales_today (int): Ventas realizadas hoy
-        - low_stock_list (array): Top 10 productos con stock crítico
-          - id (int): ID del producto
-          - name (str): Nombre
-          - sku (str): Código de barras
-          - stock (int): Cantidad actual
-    
-    Status Codes:
-        200: Estadísticas obtenidas exitosamente
-        401: No autorizado
-    """
-    
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-    
-    total_products = db.execute_query("SELECT COUNT(*) FROM items")[0][0]
-    
-    low_stock = db.execute_query(
-        "SELECT COUNT(*) FROM items WHERE quantity <= min_quantity AND quantity > 0"
-    )[0][0]
-    
-    sales_today = db.execute_query(
-        "SELECT COUNT(*) FROM sells WHERE DATE(date) = DATE('now')"
-    )[0][0]
-    
-    low_stock_items = db.execute_query(
-        "SELECT id, name, barrs_code, quantity FROM items WHERE quantity <= min_quantity ORDER BY quantity ASC LIMIT 10"
-    )
-    
-    low_stock_list = [
-        {
-            "id": row[0],
-            "name": row[1],
-            "sku": row[2],
-            "stock": row[3]
-        }
-        for row in low_stock_items
-    ]
-    
-    return jsonify({
-        "products": total_products,
-        "low_stock": low_stock,
-        "sales_today": sales_today,
-        "low_stock_list": low_stock_list
-    }), 200
-
-@api_bp.route("/sales", methods=["POST"])
-def create_sale():
-    """
-    Registra una nueva venta de un producto.
-    
-    Requiere login: True.
-    
-    Request Body (JSON):
-        barcode (str): Código de barras del producto
-        quantity (int): Cantidad a vender
-    
-    Returns:
-        JSON: Confirmación de venta
-        - message (str): Mensaje de éxito
-        - product (str): Nombre del producto vendido
-        - quantity (int): Cantidad vendida
-        - total (float): Total de la venta
-    
-    Status Codes:
-        201: Venta registrada exitosamente
-        400: Faltan campos requeridos o stock insuficiente
-        401: No autorizado
-        404: Producto no encontrado
-    """
-    
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-    
-    data = request.get_json()
-    
-    if "barcode" not in data or "quantity" not in data:
-        return jsonify({"error": "Faltan campos requeridos"}), 400
-    
-    item = db.get_item_by_barcode(data["barcode"])
-    
-    if not item:
-        return jsonify({"error": "Producto no encontrado"}), 404
-    
-    item_id, _, name, _, stock, price = item
-    qty = int(data["quantity"])
-    
-    if stock < qty:
-        return jsonify({"error": "Stock insuficiente"}), 400
-    
-    vendedor = session.get("username", "unknown")
-    payment_method = data.get("payment_method", "Efectivo")
-    db.record_product_sale(item_id, qty, vendedor, payment_method)
-    
-    return jsonify({
-        "message": f"Venta registrada: {name} x{qty}",
-        "product": name,
-        "quantity": qty,
-        "total": price * qty
-    }), 201
-
-@api_bp.route("/sales/bulk", methods=["POST"])
-def create_sales_bulk():
-    """
-    Registra una venta con múltiples productos.
-    
-    Requiere login: True.
-    
-    Request Body (JSON):
-        items (array): Lista de productos a vender
-          - item_id (int): ID del producto
-          - quantity (int): Cantidad a vender
-    
-    Returns:
-        JSON: Resultado de la operación
-        - ok (bool): True si la venta fue exitosa
-        - sale_id (int): ID de la venta creada
-        - items (array): Lista de productos vendidos con detalles
-        - total (float): Total de la venta
-    
-    Status Codes:
-        201: Venta creada exitosamente
-        400: Formato inválido o error en los datos
-        401: No autorizado
-    """
-    
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-
-    data = request.get_json(silent=True) or {}
-    items = data.get("items", [])
-    if not isinstance(items, list) or not items:
-        return jsonify({"error": "Formato inválido: items[] requerido"}), 400
-
-    validated_items = []
-    resultados = []
-    total = 0
-    
-    for idx, it in enumerate(items):
-        try:
-            item_id = int(it.get("item_id"))
-            qty = int(it.get("quantity"))
-        except (TypeError, ValueError):
-            return jsonify({"error": f"item_id/cantidad inválidos en índice {idx}"}), 400
-
-        row = db.execute_query("SELECT name, quantity, price FROM items WHERE id = ?", (item_id,))
-        if not row:
-            return jsonify({"error": f"Producto con ID {item_id} no encontrado"}), 400
-
-        name, stock, price = row[0]
-        if stock < qty:
-            return jsonify({
-                "error": "Stock insuficiente",
-                "product": name,
-                "requested": qty,
-                "available": stock
-            }), 400
-
-        validated_items.append({"item_id": item_id, "quantity": qty})
-        subtotal = round(price * qty, 2)
-        resultados.append({
-            "item_id": item_id,
-            "name": name,
-            "quantity": qty,
-            "unit_price": price,
-            "subtotal": subtotal
-        })
-        total += subtotal
-
-    user_row = db.execute_query(
-        "SELECT id, username, password, role FROM users WHERE id = ?",
-        (session.get("user_id"),)
-    )
-    vendedor = user_row[0] if user_row else None
-    payment_method = data.get("payment_method", "Efectivo")
-    try:
-        sale_id = db.record_bulk_sale(validated_items, vendedor[1], payment_method)
-    except Exception as e:
-        return jsonify({"error": str(e)}), 400
-
-    return jsonify({
-        "ok": True,
-        "sale_id": sale_id,
-        "items": resultados,
-        "total": round(total, 2)
-    }), 201
-
-@api_bp.route("/sales", methods=["GET"])
-def list_sales():
-    """
-    Lista ventas con paginacion server-side.
-
-    Requiere login: True.
-
-    Query Parameters:
-        from    (str, optional) : Fecha inicial YYYY-MM-DD
-        to      (str, optional) : Fecha final   YYYY-MM-DD
-        product (str, optional) : Filtro por nombre de producto
-        page    (int, optional) : Pagina, 1-based  (default: 1)
-        limit   (int, optional) : Ventas por pagina, max 100  (default: 10)
-
-    Returns:
-        JSON:
-            data   (list) : Ventas de la pagina actual, cada una con:
-                            id, date, products[], total_quantity, total
-            total  (int)  : Total de ventas que coinciden con los filtros
-            page   (int)  : Pagina actual
-            pages  (int)  : Total de paginas
-            limit  (int)  : Registros por pagina usados
-
-    Status Codes:
-        200: Exito
-        401: No autorizado
-    """
-
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-
-    date_from   = request.args.get("from")
-    date_to     = request.args.get("to")
-    product_q   = request.args.get("product", "").strip()
-    page        = max(1, request.args.get("page", 1, type=int))
-    limit       = min(100, max(1, request.args.get("limit", 10, type=int)))
-    offset      = (page - 1) * limit
-
-    sell_where  = ["1=1"]
-    sell_params = []
-
-    if date_from:
-        sell_where.append("DATE(s.date) >= ?")
-        sell_params.append(date_from)
-    if date_to:
-        sell_where.append("DATE(s.date) <= ?")
-        sell_params.append(date_to)
-    if product_q:
-        sell_where.append("s.id IN (SELECT d2.sell_id FROM details d2 JOIN items i2 ON d2.item_id = i2.id WHERE i2.name LIKE ?)")
-        sell_params.append(f"%{product_q}%")
-
-    sell_where_clause = " AND ".join(sell_where)
-
-    count_query = f"""
-        SELECT COUNT(DISTINCT s.id)
-        FROM sells s
-        WHERE {sell_where_clause}
-    """
-    total = db.execute_query(count_query, tuple(sell_params))[0][0]
-    pages = max(1, -(-total // limit))
-
-    ids_query = f"""
-        SELECT DISTINCT s.id
-        FROM sells s
-        WHERE {sell_where_clause}
-        ORDER BY s.date DESC, s.id DESC
-        LIMIT ? OFFSET ?
-    """
-    id_rows = db.execute_query(ids_query, tuple(sell_params) + (limit, offset))
-
-    if not id_rows:
-        return jsonify({
-            "data":  [],
-            "total": total,
-            "page":  page,
-            "pages": pages,
-            "limit": limit,
-        }), 200
-
-    sale_ids = [row[0] for row in id_rows]
-
-    placeholders = ",".join("?" * len(sale_ids))
-    detail_query = f"""
-        SELECT s.id, s.date, i.name, d.quantity, d.price, s.vendedor, s.payment_method
-        FROM sells s
-        JOIN details d ON s.id = d.sell_id
-        JOIN items  i ON d.item_id = i.id
-        WHERE s.id IN ({placeholders})
-        ORDER BY s.date DESC, s.id DESC
-    """
-    rows = db.execute_query(detail_query, tuple(sale_ids))
-
-    sales_dict = {}
-    for sale_id, date, name, quantity, price, vendedor, payment_method in rows:
-        if sale_id not in sales_dict:
-            sales_dict[sale_id] = {
-                "id":             sale_id,
-                "date":           date,
-                "products":       [],
-                "total_quantity": 0,
-                "total":          0.0,
-                "vendedor":       vendedor,
-                "payment_method": payment_method,
-            }
-        sales_dict[sale_id]["products"].append({
-            "name":     name,
-            "quantity": quantity,
-            "price":    float(price),
-        })
-        sales_dict[sale_id]["total_quantity"] += quantity
-        sales_dict[sale_id]["total"]          += quantity * float(price)
-
-    sales = []
-    for sid in sale_ids:
-        if sid in sales_dict:
-            sales_dict[sid]["total"] = round(sales_dict[sid]["total"], 2)
-            sales.append(sales_dict[sid])
-
-    return jsonify({
-        "data":  sales,
-        "total": total,
-        "page":  page,
-        "pages": pages,
-        "limit": limit,
-    }), 200
-    
-@api_bp.route("/items", methods=["GET"])
-def search_items():
-    """
-    Busca productos para autocompletado.
-    
-    Requiere login: True.
-    
-    Query Parameters:
-        q (str): Término de búsqueda (nombre o código de barras)
-    
-    Returns:
-        JSON: Lista de hasta 10 productos coincidentes
-        - id (int): ID del producto
-        - barcode (str): Código de barras
-        - name (str): Nombre
-        - description (str): Descripción
-        - stock (int): Cantidad disponible
-        - price (float): Precio
-    
-    Status Codes:
-        200: Búsqueda exitosa (puede retornar array vacío)
-        401: No autorizado
-    """
-    
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-    
-    query_param = request.args.get("q", "").strip()
-    if not query_param:
-        return jsonify([]), 200
-    
-    rows = db.execute_query(
-        "SELECT id, barrs_code, name, description, quantity, price FROM items WHERE barrs_code LIKE ? OR name LIKE ? LIMIT 10",
-        (f"%{query_param}%", f"%{query_param}%")
-    )
-    
-    items = [
-        {
-            "id": row[0],
-            "barcode": row[1],
-            "name": row[2],
-            "description": row[3],
-            "stock": row[4],
-            "price": row[5]
-        }
-        for row in rows
-    ]
-    
-    return jsonify(items), 200
-
-@api_bp.route("/sales/<int:sale_id>", methods=["GET"])
-def get_sale_detail(sale_id):
-    """
-    Obtiene los detalles de una venta específica.
-    
-    Requiere login: True.
-    
-    Args:
-        sale_id (int): ID de la venta a consultar
-    
-    Returns:
-        JSON: Detalles completos de la venta
-        - id (int): ID de la venta
-        - date (str): Fecha y hora de la venta
-        - products (array): Lista de productos vendidos
-          - name (str): Nombre del producto
-          - quantity (int): Cantidad vendida
-          - price (float): Precio unitario
-        - total (float): Total de la venta
-    
-    Status Codes:
-        200: Venta encontrada
-        401: No autorizado
-        404: Venta no encontrada
-    """
-    
-    if not session.get("user_id"):
-        return jsonify({"error": "Unauthorized"}), 401
-    
-    sale_data = db.execute_query(
-        """
-        SELECT s.id, s.date, i.name, d.quantity, d.price, s.vendedor, s.payment_method
-        FROM sells s 
-        JOIN details d ON s.id = d.sell_id 
-        JOIN items i ON d.item_id = i.id 
-        WHERE s.id = ?
-        """,
-        (sale_id,)
-    )
-    
-    if not sale_data:
-        return jsonify({"error": "Sale not found"}), 404
-    
-    sale = {
-        "id":             sale_data[0][0],
-        "date":           sale_data[0][1],
-        "products":       [],
-        "total":          0.0,
-        "vendedor":       sale_data[0][5],
-        "payment_method": sale_data[0][6],
-    }
-    
-    for row in sale_data:
-        product = {
-            "name": row[2],
-            "quantity": row[3],
-            "price": float(row[4])
-        }
-        sale["products"].append(product)
-        sale["total"] += product["quantity"] * product["price"]
-    
-    return jsonify(sale), 200, {'Content-Type': 'application/json'}
-
-@api_bp.route('/metrics', methods=['GET'])
-def get_metrics():
-    """
-    Obtiene métricas del negocio para el dashboard de analytics.
-    
-    Requiere login: True.
-    
-    Query params:
-        - period (int): Número de días (7, 30, 90, 365)
-        - from (str): Fecha inicio (YYYY-MM-DD)
-        - to (str): Fecha fin (YYYY-MM-DD)
-    
-    Returns:
-        JSON: Métricas completas del negocio
-    """
-    
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-    
-    period = request.args.get('period', 7, type=int)
-    date_from = request.args.get('from')
-    date_to = request.args.get('to')
-    
-    if date_from and date_to:
-        start_date = date_from
-        end_date = date_to
-        start_dt = datetime.strptime(start_date, '%Y-%m-%d')
-        end_dt = datetime.strptime(end_date, '%Y-%m-%d')
-        period_days = (end_dt - start_dt).days + 1
-    else:
-        end_date = datetime.now().strftime('%Y-%m-%d')
-        start_date = (datetime.now() - timedelta(days=period - 1)).strftime('%Y-%m-%d')
-        period_days = period
-    
-    start_dt = datetime.strptime(start_date, '%Y-%m-%d')
-    prev_end_date = (start_dt - timedelta(days=1)).strftime('%Y-%m-%d')
-    prev_start_date = (start_dt - timedelta(days=period_days)).strftime('%Y-%m-%d')
-    
-    data = db.get_metrics_data(start_date, end_date, prev_start_date, prev_end_date)
-    
-    revenue = float(data["kpis"][0])
-    total_sales = int(data["kpis"][1])
-    units_sold = int(data["kpis"][2])
-    avg_ticket = round(revenue / total_sales, 2) if total_sales > 0 else 0
-    
-    prev_revenue = float(data["prev_kpis"][0])
-    prev_total_sales = int(data["prev_kpis"][1])
-    prev_avg_ticket = round(prev_revenue / prev_total_sales, 2) if prev_total_sales > 0 else 0
-    prev_units_sold = int(data["prev_kpis"][2])
-    
-    def calc_change(current, previous):
-        if previous == 0:
-            return 100.0 if current > 0 else 0.0
-        return round(((current - previous) / previous) * 100, 1)
-    
-    revenue_change = calc_change(revenue, prev_revenue)
-    sales_change = calc_change(total_sales, prev_total_sales)
-    ticket_change = calc_change(avg_ticket, prev_avg_ticket)
-    units_change = calc_change(units_sold, prev_units_sold)
-    
-    date_range = {}
-    current_dt = datetime.strptime(start_date, '%Y-%m-%d')
-    end_dt = datetime.strptime(end_date, '%Y-%m-%d')
-    while current_dt <= end_dt:
-        date_range[current_dt.strftime('%Y-%m-%d')] = {"revenue": 0, "sales": 0}
-        current_dt += timedelta(days=1)
-    
-    for row in data["sales_over_time"]:
-        date_str = row[0] if isinstance(row[0], str) else row[0].strftime('%Y-%m-%d')
-        if date_str in date_range:
-            date_range[date_str] = {"revenue": float(row[1]), "sales": int(row[2])}
-    
-    days_es = ['Lun', 'Mar', 'Mié', 'Jue', 'Vie', 'Sáb', 'Dom']
-    labels, revenues, sales_counts = [], [], []
-    for date_str in sorted(date_range):
-        dt = datetime.strptime(date_str, '%Y-%m-%d')
-        labels.append(days_es[dt.weekday()] if period_days <= 7 else dt.strftime('%d/%m'))
-        revenues.append(date_range[date_str]["revenue"])
-        sales_counts.append(date_range[date_str]["sales"])
-    
-    top_products = [
-        {"id": r[0], "name": r[1], "sku": r[2] or "Sin SKU", "units": int(r[3]), "revenue": float(r[4])}
-        for r in data["top_products"]
-    ]
-    
-    sales_by_weekday = [0] * 7
-    for row in data["weekday_sales"]:
-        sqlite_wd = int(row[0])
-        adjusted = (sqlite_wd - 1) if sqlite_wd > 0 else 6
-        sales_by_weekday[adjusted] = int(row[1])
-    
-    sales_by_hour = [0] * 24
-    for row in data["hourly_sales"]:
-        sales_by_hour[int(row[0])] = int(row[1])
-    
-    out_of_stock = int(data["alerts"][0] or 0)
-    low_stock = int(data["alerts"][1] or 0)
-    no_movement = int(data["alerts"][2] or 0)
-    
-    days_names = ['Lunes', 'Martes', 'Miércoles', 'Jueves', 'Viernes', 'Sábado', 'Domingo']
-    
-    best_day = None
-    if data["best_weekday_idx"] is not None:
-        best_day = {
-            "name": days_names[data["best_weekday_idx"]],
-            "revenue": data["best_day_revenue"]
-        }
-    
-    peak_hour = sales_by_hour.index(max(sales_by_hour)) if any(sales_by_hour) else None
-    top_product = {"name": top_products[0]["name"], "units": top_products[0]["units"]} if top_products else None
-    
-    if prev_revenue > 0:
-        if revenue_change > 10:
-            trend = f"Las ventas aumentaron {revenue_change}% respecto al período anterior. ¡Excelente trabajo!"
-        elif revenue_change > 0:
-            trend = f"Las ventas aumentaron {revenue_change}% respecto al período anterior. Buen progreso."
-        elif revenue_change > -10:
-            trend = f"Las ventas disminuyeron {abs(revenue_change)}% respecto al período anterior. Considera revisar tu estrategia."
-        else:
-            trend = f"Las ventas cayeron {abs(revenue_change)}% respecto al período anterior. Se recomienda tomar acción."
-    else:
-        trend = "No hay datos del período anterior para comparar."
-    
-    return jsonify({
-        "kpis": {
-            "revenue": round(revenue, 2),
-            "totalSales": total_sales,
-            "avgTicket": avg_ticket,
-            "unitsSold": units_sold,
-            "revenueChange": revenue_change,
-            "salesChange": sales_change,
-            "ticketChange": ticket_change,
-            "unitsChange": units_change
-        },
-        "salesOverTime": {
-            "labels": labels,
-            "revenue": revenues,
-            "sales": sales_counts
-        },
-        "topProducts": top_products,
-        "salesByWeekday": sales_by_weekday,
-        "salesByHour": sales_by_hour,
-        "comparison": {
-            "current": round(revenue, 2),
-            "previous": round(prev_revenue, 2)
-        },
-        "alerts": {
-            "outOfStock": out_of_stock,
-            "lowStock": low_stock,
-            "noMovement": no_movement
-        },
-        "insights": {
-            "bestDay": best_day,
-            "peakHour": peak_hour,
-            "topProduct": top_product,
-            "trend": trend
-        },
-        "period": {
-            "start": start_date,
-            "end": end_date,
-            "days": period_days
-        }
-    }), 200
-    
-def generate_reset_code():
-    """Genera un código de recuperación aleatorio de 6 dígitos."""
-    import secrets
-    return f"{secrets.randbelow(10**6):06d}"
-
-@api_bp.route("/users/reset-password", methods=["POST"])
-def restore_password():
-    """
-    Endpoint para restaurar la contraseña de un usuario.
-    
-    Requiere login: False.
-    
-    Request Body (JSON):
-        email (str): Correo electrónico del usuario
-    
-    Returns:
-        JSON: {"message": "Codigo enviado al correo"}
-    
-    Status Codes:
-        200: Codigo enviado exitosamente
-        400: Faltan campos requeridos o formato inválido
-        404: Usuario no encontrado
-    
-    Example Request:
-        POST /users/reset-password
-        {
-            "email": "usuario@ejemplo.com"
-        }
-    """
-    
-    data = request.get_json()
-    
-    logger.info(f"Password reset requested for email: {data.get('email', 'N/A')}")
-    
-    if "email" not in data:
-        return jsonify({"error": "Faltan campos requeridos"}), 400
-    
-    email = data["email"].strip()
-    
-    if not UserValidator.validate_email(email):
-        return jsonify({"error": "Formato de correo inválido"}), 400
-    
-    user = db.get_user_by_email(email)
-    
-    if not user:
-        return jsonify({"error": "Usuario no encontrado"}), 404
-    
-    reset_code = generate_reset_code()
-    db.save_reset_code(email, reset_code)
-    
-    email_sender.send_email(
-        email,
-        subject="Código de recuperación de contraseña",
-        body=f"Tu código de recuperación es: {reset_code}. Este código es válido por 15 minutos."
-    )
-    
-    return jsonify({"message": "Codigo enviado al correo"}), 200
-
-@api_bp.route("/users/validate-code", methods=["POST"])
-def verify_code():
-    """
-    Endpoint para verificar el código de recuperación de contraseña.
-    
-    Requiere login: False.
-    
-    Request Body (JSON):
-        email (str): Correo electrónico del usuario
-        code (str): Código de recuperación enviado al correo
-    
-    Returns:
-        JSON: {"message": "Código verificado, puedes restablecer tu contraseña"}
-    
-    Status Codes:
-        200: Código verificado exitosamente
-        400: Faltan campos requeridos o formato inválido
-        404: Usuario no encontrado
-        401: Código inválido o expirado
-    
-    Example Request:
-        POST /users/validate-code
-        {
-            "email": "usuario@ejemplo.com",
-            "code": "123456"
-        }
-    """
-    data = request.get_json()
-    
-    if "email" not in data or "code" not in data:
-        return jsonify({"error": "Faltan campos requeridos"}), 400
-    
-    code_status = db.verify_code(email=data.get("email", ""), code=data.get("code", ""))
-
-    if not code_status:
-        return jsonify({"error": "Código inválido o expirado"}), 401
-    
-    return jsonify({"message": "Código verificado, puedes restablecer tu contraseña"}), 200
-
-@api_bp.route("/users/reset-password/change-password", methods=["POST"])
-def change_password():
-    """
-    Endpoint para cambiar la contraseña después de verificar el código de recuperación.
-    
-    Requiere login: False.
-    
-    Request Body (JSON):
-        email (str): Correo electrónico del usuario
-        new_password (str): Nueva contraseña a establecer
-    
-    Returns:
-        JSON: {"message": "Contraseña restablecida exitosamente"}
-    
-    Status Codes:
-        200: Contraseña cambiada exitosamente
-        400: Faltan campos requeridos o formato inválido
-        404: Usuario no encontrado
-        401: No autorizado (si el código no fue verificado)
-    
-    Example Request:
-        POST /users/reset-password/change-password
-        {
-            "email": "usuario@ejemplo.com",
-            "code": "123456",
-            "new_password": "nueva_contraseña_segura"
-        }
-    """
-    data = request.get_json()
-    
-    if not db.verify_code(email=data.get("email", ""), code=data.get("code", "")):
-        return jsonify({"error": "Código no verificado o expirado"}), 401
-    
-    if "email" not in data or "new_password" not in data:
-        return jsonify({"error": "Faltan campos requeridos"}), 400
-    
-    email = data["email"].strip()
-    new_password = data["new_password"].strip()
-    
-    if not UserValidator.validate_email(email):
-        return jsonify({"error": "Formato de correo inválido"}), 400
-    
-    user = db.get_user_by_email(email)
-    
-    if not user:
-        return jsonify({"error": "Usuario no encontrado"}), 404
-    
-    hashed = generate_password_hash(new_password)
-    db.update_user_password(email, hashed)
-    
-    db.delete_reset_code(email=data["email"])
-    
-    return jsonify({"message": "Contraseña restablecida exitosamente"}), 200
-
-@api_bp.route("/users", methods=["GET"])
-def get_users():
-    """Lista todos los usuarios con paginación."""
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-
-    if session.get("role") != "admin":
-        return jsonify({"error": "Permiso denegado"}), 403
-
-    search = request.args.get("search", "").strip()
-    page   = max(1, request.args.get("page", 1, type=int))
-    limit  = min(100, max(1, request.args.get("limit", 10, type=int)))
-    offset = (page - 1) * limit
-
-    where  = ["1=1"]
-    params = []
-
-    if search:
-        where.append("(username LIKE ? OR email LIKE ?)")
-        params.extend([f"%{search}%", f"%{search}%"])
-
-    where_clause = " AND ".join(where)
-
-    total = db.execute_query(
-        f"SELECT COUNT(*) FROM users WHERE {where_clause}", tuple(params)
-    )[0][0]
-    pages = max(1, -(-total // limit))
-
-    rows = db.execute_query(
-        f"""
-        SELECT id, username, email, role, status, created_at
-        FROM users
-        WHERE {where_clause}
-        ORDER BY created_at DESC
-        LIMIT ? OFFSET ?
-        """,
-        tuple(params) + (limit, offset)
-    )
-
-    users_list = [
-        {
-            "id":         row[0],
-            "username":   row[1],
-            "email":      row[2],
-            "role":       row[3],
-            "status":     row[4],
-            "created_at": row[5],
-        }
-        for row in rows
-    ]
-
-    return jsonify({
-        "data":  users_list,
-        "total": total,
-        "page":  page,
-        "pages": pages,
-        "limit": limit,
-    }), 200
-
-
-@api_bp.route("/users/<int:user_id>", methods=["GET"])
-def get_user(user_id):
-    """Obtiene un usuario específico."""
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-
-    if session.get("role") != "admin":
-        return jsonify({"error": "Permiso denegado"}), 403
-
-    rows = db.execute_query(
-        "SELECT id, username, email, role, status, created_at FROM users WHERE id = ?",
-        (user_id,)
-    )
-    if not rows:
-        return jsonify({"error": "Usuario no encontrado"}), 404
-
-    r = rows[0]
-    return jsonify({
-        "id":         r[0],
-        "username":   r[1],
-        "email":      r[2],
-        "role":       r[3],
-        "status":     r[4],
-        "created_at": r[5],
-    }), 200
-
-
-@api_bp.route("/users", methods=["POST"])
-def create_user():
-    """Crea un nuevo usuario."""
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-
-    if session.get("role") != "admin":
-        return jsonify({"error": "Permiso denegado"}), 403
-
-    data = request.get_json()
-    required = ["username", "email", "password", "role"]
-    if not all(f in data for f in required):
-        return jsonify({"error": "Faltan campos requeridos"}), 400
-
-    try:
-        UserValidator.validate_email(data["email"])
-    except Exception:
-        return jsonify({"error": "Formato de correo inválido"}), 400
-
-    existing = db.get_user_by_email(data["email"].strip())
-    if existing:
-        return jsonify({"error": "El correo ya está registrado"}), 409
-
-    hashed = generate_password_hash(data["password"])
-    try:
-        db.execute_query(
-            "INSERT INTO users (username, email, password, role, status) VALUES (?, ?, ?, ?, 1)",
-            (data["username"].strip(), data["email"].strip(), hashed, data["role"]),
-            fetch=False
-        )
-        logger.info(f"Usuario creado: {data['username']} por admin ID {session.get('user_id')}")
-        return jsonify({"message": "Usuario creado exitosamente"}), 201
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
-
-@api_bp.route("/users/<int:user_id>", methods=["PUT"])
-def update_user(user_id):
-    """Actualiza un usuario existente."""
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-
-    if session.get("role") != "admin":
-        return jsonify({"error": "Permiso denegado"}), 403
-
-    data = request.get_json()
-    updates = []
-    params  = []
-
-    if "username" in data:
-        updates.append("username = ?")
-        params.append(data["username"].strip())
-    if "email" in data:
-        updates.append("email = ?")
-        params.append(data["email"].strip())
-    if "role" in data:
-        updates.append("role = ?")
-        params.append(data["role"])
-    if "status" in data:
-        updates.append("status = ?")
-        params.append(int(data["status"]))
-    if "password" in data and data["password"]:
-        updates.append("password = ?")
-        params.append(generate_password_hash(data["password"]))
-
-    if not updates:
-        return jsonify({"error": "No hay datos para actualizar"}), 400
-
-    params.append(user_id)
-    db.execute_query(
-        f"UPDATE users SET {', '.join(updates)} WHERE id = ?",
-        tuple(params),
-        fetch=False
-    )
-    logger.info(f"Usuario ID {user_id} actualizado por admin ID {session.get('user_id')}")
-    return jsonify({"message": "Usuario actualizado"}), 200
-
-
-@api_bp.route("/users/<int:user_id>", methods=["DELETE"])
-def delete_user(user_id):
-    """Deshabilita un usuario (baja lógica)."""
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-
-    if session.get("role") != "admin":
-        return jsonify({"error": "Permiso denegado"}), 403
-
-    if user_id == session.get("user_id"):
-        return jsonify({"error": "No puedes darte de baja a ti mismo"}), 400
-
-    db.execute_query(
-        "UPDATE users SET status = 0 WHERE id = ?",
-        (user_id,),
-        fetch=False
-    )
-    logger.info(f"Usuario ID {user_id} dado de baja por admin ID {session.get('user_id')}")
-    return jsonify({"message": "Usuario dado de baja"}), 200
-
-
-@api_bp.route("/users/<int:user_id>/activity", methods=["GET"])
-def get_user_activity(user_id):
-    """Obtiene los movimientos/ventas de un usuario."""
-    auth_error = require_auth()
-    if auth_error:
-        return auth_error
-
-    if session.get("role") != "admin":
-        return jsonify({"error": "Permiso denegado"}), 403
-
-    page   = max(1, request.args.get("page", 1, type=int))
-    limit  = min(100, max(1, request.args.get("limit", 10, type=int)))
-    offset = (page - 1) * limit
-
-    user_rows = db.execute_query(
-        "SELECT username FROM users WHERE id = ?", (user_id,)
-    )
-    if not user_rows:
-        return jsonify({"error": "Usuario no encontrado"}), 404
-
-    username = user_rows[0][0]
-
-    total = db.execute_query(
-        "SELECT COUNT(*) FROM sells WHERE vendedor = ?", (username,)
-    )[0][0]
-    pages = max(1, -(-total // limit))
-
-    rows = db.execute_query(
-        """
-        SELECT s.id, s.date, s.payment_method,
-               COUNT(d.id) AS items,
-               SUM(d.quantity * d.price) AS total
-        FROM sells s
-        JOIN details d ON s.id = d.sell_id
-        WHERE s.vendedor = ?
-        GROUP BY s.id
-        ORDER BY s.date DESC
-        LIMIT ? OFFSET ?
-        """,
-        (username, limit, offset)
-    )
-
-    activity = [
-        {
-            "sale_id":        r[0],
-            "date":           r[1],
-            "payment_method": r[2],
-            "items":          int(r[3]),
-            "total":          round(float(r[4]), 2),
-        }
-        for r in rows
-    ]
-
-    return jsonify({
-        "username": username,
-        "data":     activity,
-        "total":    total,
-        "page":     page,
-        "pages":    pages,
-        "limit":    limit,
-    }), 200
