+++ api/__init__.py
@@ -1,0 +1,15 @@
+from flask import Blueprint
+
+api_bp = Blueprint("api", __name__)
+
+from api.products_api import products_api
+from api.sales_api import sales_api
+from api.users_api import users_api
+from api.metrics_api import metrics_api
+from api.health_api import health_api
+
+api_bp.register_blueprint(products_api)
+api_bp.register_blueprint(sales_api)
+api_bp.register_blueprint(users_api)
+api_bp.register_blueprint(metrics_api)
+api_bp.register_blueprint(health_api)
