+++ templates/404.html
@@ -5,7 +5,7 @@
   <h1 class="display-5">Error 404</h1>
   <p class="lead mb-4" id="phrase">Esta página se ha ido a hacer inventario.</p>
 
-  <a href="{{ url_for('index') }}" class="btn btn-primary btn-sm">Volver al Dashboard</a>
+  <a href="{{ url_for('dashboard.index') }}" class="btn btn-primary btn-sm">Volver al Dashboard</a>
   <a href="javascript:history.back()" class="btn btn-outline-secondary btn-sm">Regresar</a>
 </div>
 {% endblock %}
