+++ templates/reset_password.html
@@ -240,7 +240,7 @@
         </form>
 
         <div class="back-link">
-          <a href="{{ url_for('login') }}">← Volver al inicio de sesión</a>
+          <a href="{{ url_for('auth.login') }}">← Volver al inicio de sesión</a>
         </div>
       </div>
     </div>
