+++ templates/under_development.html
@@ -21,7 +21,7 @@
       <span class="progress-text text-muted">En progreso...</span>
     </div>
     <div class="dev-actions">
-      <a href="{{ url_for('index') }}" class="btn">Volver al Dashboard</a>
+      <a href="{{ url_for('dashboard.index') }}" class="btn">Volver al Dashboard</a>
     </div>
   </div>
 </div>
